test: additional mock implementations
Implements additional mocks to allow for improved testing.
- `SimpleOutboxType`: An `OutboxType` where its type is specified
  during construction. Accordingly, `DummyOutboxType` is updated to
  extend it and simply specify a fixed type.
- `MockOutboxHandler`: An `OutboxHandler` where the supported type and
  (optionally) the used `Clock` are specified during construction.
  The rest of the implementation is copied from `DummyOutboxHandler`
  which is now updated to extend the new class with a fixed type.

[SW-587]
