+++ core/src/test/groovy/io/github/bluegroundltd/outbox/utils/DummyOutboxType.groovy
@@ -1,0 +1,10 @@
+package io.github.bluegroundltd.outbox.utils
+
+import groovy.transform.EqualsAndHashCode
+
+@EqualsAndHashCode(includeFields=true)
+class DummyOutboxType extends SimpleOutboxType {
+  DummyOutboxType() {
+    super("dummyOutboxType")
+  }
+}
