+++ core/src/test/groovy/io/github/bluegroundltd/outbox/utils/OutboxItemBuilder.groovy
@@ -1,6 +1,5 @@
 package io.github.bluegroundltd.outbox.utils
 
-import groovy.transform.EqualsAndHashCode
 import io.github.bluegroundltd.outbox.item.OutboxItem
 import io.github.bluegroundltd.outbox.item.OutboxStatus
 import io.github.bluegroundltd.outbox.item.OutboxType
@@ -87,13 +86,5 @@ class OutboxItemBuilder implements SpecHelper {
       rerunAfter,
       deleteAfter
     )
-  }
-}
-
-@EqualsAndHashCode(includeFields=true)
-class DummyOutboxType implements OutboxType {
-  @Override
-  String getType() {
-    return "dummyOutboxType"
   }
 }
